Refactor post draft component Re-styled ReadOnly channels Remove filename from upload error
