+++ .eslintrc.json
@@ -44,7 +44,8 @@
     "@typescript-eslint/no-explicit-any": "warn",
     "@typescript-eslint/no-use-before-define": 0,
     "@typescript-eslint/no-var-requires": 0,
-    "@typescript-eslint/explicit-function-return-type": 0
+    "@typescript-eslint/explicit-function-return-type": 0,
+    "@typescript-eslint/explicit-module-boundary-types": "off"
   },
   "overrides": [
     {
