+++ app/components/post_draft/quick_actions/camera_quick_action/camera_quick_action.test.js
@@ -18,7 +18,6 @@ jest.mock('react-native-image-picker', () => ({
 describe('CameraButton', () => {
     const formatMessage = jest.fn();
     const baseProps = {
-        blurTextBox: jest.fn(),
         fileCount: 0,
         maxFileCount: 5,
         onShowFileMaxWarning: jest.fn(),
