+++ app/components/post_draft/quick_actions/file_quick_action/file_quick_action.test.js
@@ -18,7 +18,6 @@ jest.mock('react-native-image-picker', () => ({
 describe('FileQuickAction', () => {
     const formatMessage = jest.fn();
     const baseProps = {
-        blurTextBox: jest.fn(),
         fileCount: 0,
         maxFileCount: 5,
         onShowFileMaxWarning: jest.fn(),
