+++ app/components/post_draft/quick_actions/quick_actions.js
@@ -5,36 +5,27 @@ import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 import {Platform, StyleSheet, View} from 'react-native';
 
-import {MAX_FILE_COUNT} from '@constants/post_draft';
+import {MAX_FILE_COUNT, UPLOAD_FILES} from '@constants/post_draft';
 import EventEmitter from '@mm-redux/utils/event_emitter';
 
 import CameraAction from './camera_quick_action';
 import ImageAction from './image_quick_action';
 import FileAction from './file_quick_action';
 import InputAction from './input_quick_action';
-import SendAction from './send_action';
 
 export default class QuickActions extends PureComponent {
     static propTypes = {
-        blurTextBox: PropTypes.func.isRequired,
         canUploadFiles: PropTypes.bool,
-        canSend: PropTypes.bool,
         fileCount: PropTypes.number,
-        initialValue: PropTypes.string,
         inputEventType: PropTypes.string.isRequired,
         maxFileSize: PropTypes.number.isRequired,
-        onSend: PropTypes.func.isRequired,
-        onShowFileMaxWarning: PropTypes.func.isRequired,
         onTextChange: PropTypes.func.isRequired,
-        onUploadFiles: PropTypes.func.isRequired,
-        readonly: PropTypes.bool,
         theme: PropTypes.object.isRequired,
     };
 
     static defaultProps = {
         canUploadFiles: true,
         fileCount: 0,
-        initialValue: '',
     };
 
     constructor(props) {
@@ -42,8 +33,8 @@ export default class QuickActions extends PureComponent {
 
         this.state = {
             inputValue: '',
-            atDisabled: props.readonly,
-            slashDisabled: props.readonly,
+            atDisabled: false,
+            slashDisabled: false,
         };
     }
 
@@ -51,76 +42,58 @@ export default class QuickActions extends PureComponent {
         EventEmitter.on(this.props.inputEventType, this.handleInputEvent);
     }
 
-    componentDidUpdate(prevProps) {
-        if (prevProps.readonly !== this.props.readonly || prevProps.initialValue !== this.props.initialValue) {
-            this.handleInputEvent(this.props.initialValue);
-        }
-    }
-
     componentWillUnmount() {
         EventEmitter.off(this.props.inputEventType, this.handleInputEvent);
     }
 
     handleInputEvent = (inputValue) => {
-        const {readonly} = this.props;
-        const atDisabled = readonly || inputValue[inputValue.length - 1] === '@';
-        const slashDisabled = readonly || inputValue.length > 0;
+        const atDisabled = inputValue[inputValue.length - 1] === '@';
+        const slashDisabled = inputValue.length > 0;
 
         this.setState({atDisabled, slashDisabled, inputValue});
     };
 
-    onShowFileMaxWarning = () => {
-        EventEmitter.emit('fileMaxWarning');
-    };
+    handleOnTextChange = (newValue) => {
+        this.handleInputEvent(newValue);
+        this.props.onTextChange(newValue);
+    }
+
+    handleUploadFiles(files) {
+        EventEmitter.emit(UPLOAD_FILES, files);
+    }
 
     render() {
         const {
-            blurTextBox,
             canUploadFiles,
-            canSend,
             fileCount,
-            onSend,
-            onTextChange,
-            onShowFileMaxWarning,
-            readonly,
             theme,
-            onUploadFiles,
         } = this.props;
         const uploadProps = {
-            blurTextBox,
-            disabled: !canUploadFiles || readonly,
+            disabled: !canUploadFiles,
             fileCount,
             maxFileCount: MAX_FILE_COUNT,
-            onShowFileMaxWarning,
             theme,
-            onUploadFiles,
+            onUploadFiles: this.handleUploadFiles,
         };
 
         return (
-            <View style={style.container}>
-                <View style={style.quickActionsContainer}>
-                    <InputAction
-                        disabled={this.state.atDisabled}
-                        inputType='at'
-                        onTextChange={onTextChange}
-                        theme={theme}
-                        value={this.state.inputValue}
-                    />
-                    <InputAction
-                        disabled={this.state.slashDisabled}
-                        inputType='slash'
-                        onTextChange={onTextChange}
-                        theme={theme}
-                    />
-                    <FileAction {...uploadProps}/>
-                    <ImageAction {...uploadProps}/>
-                    <CameraAction {...uploadProps}/>
-                </View>
-                <SendAction
-                    disabled={!canSend}
-                    handleSendMessage={onSend}
+            <View style={style.quickActionsContainer}>
+                <InputAction
+                    disabled={this.state.atDisabled}
+                    inputType='at'
+                    onTextChange={this.handleOnTextChange}
+                    theme={theme}
+                    value={this.state.inputValue}
+                />
+                <InputAction
+                    disabled={this.state.slashDisabled}
+                    inputType='slash'
+                    onTextChange={this.handleOnTextChange}
                     theme={theme}
                 />
+                <FileAction {...uploadProps}/>
+                <ImageAction {...uploadProps}/>
+                <CameraAction {...uploadProps}/>
             </View>
         );
     }
