+++ app/components/post_draft/send_action/paper_plane.js
