+++ assets/base/i18n/en.json
@@ -79,7 +79,7 @@
   "edit_post.editPost": "Edit the post...",
   "edit_post.save": "Save",
   "file_attachment.download": "Download",
-  "file_upload.fileAbove": "File above {max} cannot be uploaded: {filename}",
+  "file_upload.fileAbove": "File above {max} cannot be uploaded",
   "get_post_link_modal.title": "Copy Link",
   "integrations.add": "Add",
   "intro_messages.anyMember": " Any member can join and read this channel.",
